+++ src/extension/commands/updateBruinCLI.ts
@@ -1,8 +1,7 @@
 import { BruinInstallCLI } from "../../bruin/bruinInstallCli";
-import { checkBruinCliInstallation } from "../../bruin";
 
 export const installOrUpdateCli = async () => {
   const bruinInstaller = new BruinInstallCLI();
-  const isInstalled = await checkBruinCliInstallation();
+  const isInstalled = await bruinInstaller.checkBruinCliInstallation();
   await bruinInstaller.installOrUpdate(isInstalled.installed);
 };